feat: 해당 박스 외부 클릭 시 닫히는 기능 커스텀 훅으로 빼기
